+++ src/js/toggleLoader.js
